+++ entrypoint-monitoring/versions.tf
@@ -4,9 +4,8 @@ terraform {
 
   required_providers {
     aws = {
-      source  = "hashicorp/aws"
-      # Needed for lambda module
-      version = ">= 4.8.0"
+      source = "hashicorp/aws"
+      version = ">= 5.26.0"
     }
     archive = {
       source  = "hashicorp/archive"