+++ webhook-converters/github-to-teams/versions.tf
@@ -5,8 +5,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      # Needed for lambda module
-      version = ">= 4.8.0"
+      version = ">= 5.26.0"
     }
     archive = {
       source  = "hashicorp/archive"